backend/requirements.txt
django>=4
graphene-django>=3
django-graphql-jwt>=0.3.4
graphene
django_apscheduler
gunicorn
whitenoise>=6.4.0
django-on-heroku
dj_database_url
python-dotenv
bidict
boto3
django-storages
graphene-file-upload
Pillow
qrcode
django-admin-interface
fpdf2
pytz
matplotlib
IPython
black
Pillow
django-dbbackup
gnupg